Remove LaunchedEffect from CountryInfoScreen
In my testing on the simulator, there's always enough time to see the loading screen!

But for Refresh, I can't see it since the VM updates immediately with the cached (mock) data.

So I'm adding in a temporary delay there and left a TODO to remind myself to completely remove it later.